dont specify name twice in processfile

+++ dib2cloud/app.py
@@ -34,9 +34,9 @@ class DibProcess(object):
         with open(path, 'r') as fh:
             return DibProcess(**yaml.safe_load(fh))
 
-    def __init__(self, name, log_dir, pf_dir, images_dir,
+    def __init__(self, log_dir, pf_dir, images_dir,
                  image_config, uuid, output_formats, pid=None):
-        self.name = name
+        self.name = image_config['name']
         self.log_dir = log_dir
         self.pf_dir = pf_dir
         self.images_dir = images_dir
@@ -82,7 +82,7 @@ class DibProcess(object):
     def to_yaml_file(self, path):
         with open(path, 'w') as fh:
             out = {}
-            for attr in ('name', 'log_dir', 'pf_dir', 'images_dir',
+            for attr in ('log_dir', 'pf_dir', 'images_dir',
                          'image_config', 'uuid', 'output_formats', 'pid'):
                 out[attr] = getattr(self, attr)
             yaml.safe_dump(out, fh)
@@ -135,8 +135,7 @@ class App(object):
         # TODO(greghaynes) determine output_formats based on provider
         output_formats = ['qcow2']
 
-        process = DibProcess(name,
-                             self.config['buildlog_dir'],
+        process = DibProcess(self.config['buildlog_dir'],
                              self.config['processfile_dir'],
                              self.config['images_dir'],
                              self.config.get_diskimage_by_name(name),
@@ -154,6 +153,9 @@ class App(object):
         if not os.path.exists(pf_path):
             raise ValueError('No build with id %s found' % image_id)
         build = DibProcess.from_processfile(pf_path)
+        if build.is_running():
+            raise ValueError('Cannot delete build %s while it is running' %
+                             image_id)
         for path in build.dest_paths:
             try:
                 os.unlink(path)
